// frontend/src/app/models/eventsource.model.ts
export class Eventsource {
    id?: any;
  business_key?: string;
  priority?: string;
  source_bu?: string;
  dcp_reference?: string;
  created_by?: string;
  updated_by?: string;
  account_name?: string;
  transaction_currency?: string;
  transaction_amount?: number;
  locked_by?: string;
  debit_account_number?: string;
  account_currency?: string;
  beneficiary_name?: string;
  cust_info_mkr?: string;
  account_info_mkr?: string;
  outcome?: string;
  extension?: string;
  contact_person?: string;
  customer_called_on?: string;
  comment?: string;
  created_on?: Date;
  updated_on?: Date;
  status?: string;
}